fix: remove duplicates and move structure to correct directory
