+++ lib/features/digital_guide/tabs/structure/data/models/corridor.dart
