+++ lib/features/digital_guide/tabs/structure/data/models/information_point.dart
