+++ lib/features/digital_guide/tabs/structure/data/models/ramp.dart
