+++ lib/features/digital_guide/tabs/structure/data/repository/information_points_repository.dart
@@ -1,7 +1,7 @@
 import "package:fast_immutable_collections/fast_immutable_collections.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
 import "package:riverpod_annotation/riverpod_annotation.dart";
-import "../../../../../digital_guide/data/api/digital_guide_get_and_cache.dart";
+import "../../../../data/api/digital_guide_get_and_cache.dart";
 import "../models/information_point.dart";
 
 part "information_points_repository.g.dart";
