+++ lib/features/digital_guide/tabs/structure/data/repository/region_rooms_repository.dart
@@ -2,8 +2,8 @@ import "package:fast_immutable_collections/fast_immutable_collections.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
 import "package:riverpod_annotation/riverpod_annotation.dart";
 
-import "../../../../../digital_guide/data/api/digital_guide_get_and_cache.dart";
-import "../../../../../digital_guide/tabs/rooms/data/models/digital_guide_room.dart";
+import "../../../../data/api/digital_guide_get_and_cache.dart";
+import "../../../rooms/data/models/digital_guide_room.dart";
 
 part "region_rooms_repository.g.dart";
