+++ lib/features/digital_guide/tabs/structure/data/repository/stairways_repository.dart
@@ -2,7 +2,7 @@ import "package:fast_immutable_collections/fast_immutable_collections.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
 import "package:riverpod_annotation/riverpod_annotation.dart";
 
-import "../../../../../digital_guide/data/api/digital_guide_get_and_cache.dart";
+import "../../../../data/api/digital_guide_get_and_cache.dart";
 import "../models/stairway.dart";
 
 part "stairways_repository.g.dart";
