+++ lib/features/digital_guide/tabs/structure/domain/digital_guide_region_use_cases.dart
@@ -1,14 +1,14 @@
 import "package:fast_immutable_collections/fast_immutable_collections.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
 import "package:riverpod_annotation/riverpod_annotation.dart";
-import "../../../../digital_guide/data/models/region.dart";
-import "../../../../digital_guide/tabs/dressing_room/data/models/digital_guide_dressing_room.dart";
-import "../../../../digital_guide/tabs/dressing_room/data/repository/dressing_rooms_repository.dart";
-import "../../../../digital_guide/tabs/lifts/data/models/digital_guide_lift.dart";
-import "../../../../digital_guide/tabs/lifts/data/repository/lifts_repository.dart";
-import "../../../../digital_guide/tabs/lodge/data/models/digital_guide_lodge.dart";
-import "../../../../digital_guide/tabs/lodge/data/repository/lodges_repository.dart";
-import "../../../../digital_guide/tabs/rooms/data/models/digital_guide_room.dart";
+import "../../../data/models/region.dart";
+import "../../dressing_room/data/models/digital_guide_dressing_room.dart";
+import "../../dressing_room/data/repository/dressing_rooms_repository.dart";
+import "../../lifts/data/models/digital_guide_lift.dart";
+import "../../lifts/data/repository/lifts_repository.dart";
+import "../../lodge/data/models/digital_guide_lodge.dart";
+import "../../lodge/data/repository/lodges_repository.dart";
+import "../../rooms/data/models/digital_guide_room.dart";
 import "../data/models/corridor.dart";
 import "../data/models/information_point.dart";
 import "../data/models/parking.dart";
