+++ lib/features/digital_guide/tabs/structure/presentation/structure_expansion_tile_content.dart
@@ -4,10 +4,10 @@ import "package:flutter_riverpod/flutter_riverpod.dart";
 
 import "../../../../../config/ui_config.dart";
 import "../../../../../widgets/my_error_widget.dart";
-import "../../../../digital_guide/data/models/digital_guide_response.dart";
-import "../../../../digital_guide/data/models/level_with_regions.dart";
-import "../../../../digital_guide/data/repository/levels_repository.dart";
-import "../../../../digital_guide/presentation/widgets/digital_guide_nav_link.dart";
+import "../../../data/models/digital_guide_response.dart";
+import "../../../data/models/level_with_regions.dart";
+import "../../../data/repository/levels_repository.dart";
+import "../../../presentation/widgets/digital_guide_nav_link.dart";
 import "../../../../navigator/utils/navigation_commands.dart";
 
 class StructureExpansionTileContent extends ConsumerWidget {
