+++ lib/features/digital_guide/tabs/structure/presentation/views/corridor_view.dart
@@ -7,9 +7,9 @@ import "../../../../../../config/ui_config.dart";
 import "../../../../../../theme/app_theme.dart";
 import "../../../../../../utils/context_extensions.dart";
 import "../../../../../../widgets/detail_views/detail_view_app_bar.dart";
-import "../../../../../digital_guide/presentation/widgets/bullet_list.dart";
-import "../../../../../digital_guide/presentation/widgets/digital_guide_image.dart";
-import "../../../../../digital_guide/presentation/widgets/digital_guide_nav_link.dart";
+import "../../../../presentation/widgets/bullet_list.dart";
+import "../../../../presentation/widgets/digital_guide_image.dart";
+import "../../../../presentation/widgets/digital_guide_nav_link.dart";
 import "../../data/models/corridor.dart";
 
 @RoutePage()
