+++ lib/features/navigator/utils/navigation_commands.dart
@@ -13,7 +13,7 @@ import "../../digital_guide/tabs/adapted_toilets/data/models/adapted_toilet.dart
 import "../../digital_guide/tabs/lifts/data/models/digital_guide_lift.dart";
 import "../../digital_guide/tabs/micronavigation/data/models/micronavigation_response.dart";
 import "../../digital_guide/tabs/rooms/data/models/digital_guide_room.dart";
-import "../../digital_guide_view/tabs/structure/data/models/corridor.dart";
+import "../../digital_guide/tabs/structure/data/models/corridor.dart";
 import "../../digital_guide_view/tabs/transportation/data/models/digital_guide_transportation.dart";
 import "../../parkings/parkings_view/models/parking.dart";
 import "../app_router.dart";
